+++ js/utils.js
@@ -1,0 +1,15 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  let previousResult = -1;
+  return () => {
+    const result = Math.floor(Math.random() * (upper - lower + 1) + lower);
+    if (previousResult !== result) {
+      previousResult = result;
+      return result;
+    }
+    return result === upper ? lower : result + 1;
+  };
+};
+
+export {getRandomInteger};
